requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.6.1
xgboost==2.0.2
lightgbm==4.1.0
catboost==1.2.2
yfinance==0.2.33
requests==2.31.0
fredapi==0.5.1
joblib==1.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
redis==5.0.1
slowapi==0.1.9
psycopg2-binary==2.9.9
aiohttp==3.9.1